+++ packages/x-adapter-platform/src/endpoint-adapters/utils.ts
@@ -41,12 +41,12 @@ export function getBeaconServiceUrl(from: ExtraParamsRequest): string {
  */
 export function getAiSuggestionsServiceUrl(from: ExtraParamsRequest): string {
   if (from.extraParams?.env === 'test') {
-    return 'https://questions.internal.test.empathy.co/overview'
+    return 'https://questions.internal.test.empathy.co/suggestions'
   }
   if (from.extraParams?.env === 'staging') {
-    return 'https://questions.staging.empathy.co/v1/overview'
+    return 'https://questions.staging.empathy.co/v1/suggestions'
   }
-  return 'https://api.{extraParams.env(.)}empathy.co/v1/overview'
+  return 'https://api.{extraParams.env(.)}empathy.co/v1/suggestions'
 }
 
 /**